+++ src/components/navbar/navbar.tsx
@@ -1,12 +1,7 @@
 import React, { useState } from "react";
 import MenuIcon from "@mui/icons-material/Menu";
 import LogoutIcon from "@mui/icons-material/Logout";
-import { AccountCircle } from "@mui/icons-material";
-import { signOut, useSession } from "next-auth/react";
-
 import {
-  AppBar,
-  Avatar,
   Box,
   CssBaseline,
   Divider,
@@ -15,81 +10,16 @@ import {
   Toolbar,
   Typography,
   Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
+  AppBar,
 } from "@mui/material";
+import ProfilePopup from "../pop-up/ProfilePopup";
+import { signOut, useSession } from "next-auth/react";
 
 interface Props {
   window?: () => Window;
 }
 
 const drawerWidth = 240;
-
-function PopUpButton({ userData }: { userData: any }) {
-  const [open, setOpen] = useState(false);
-
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <>
-      <IconButton
-        size="large"
-        aria-haspopup="true"
-        color="inherit"
-        onClick={handleClickOpen}
-      >
-        <AccountCircle />
-      </IconButton>
-
-      <Dialog fullWidth maxWidth="xs" open={open} onClose={handleClose}>
-        <DialogContent>
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "center",
-            }}
-          >
-            <DialogTitle>Your Profile</DialogTitle>
-            <Avatar
-              src={userData.image}
-              sx={{
-                width: 100,
-                height: 100,
-                mb: 2,
-              }}
-            />
-            <Typography>
-              Name: {userData.name ? userData.name : "Alien"}
-            </Typography>
-            <Typography>Email: {userData.email}</Typography>
-            <Button
-              variant="outlined"
-              startIcon={<LogoutIcon />}
-              onClick={() => signOut()}
-              sx={{ mt: 2 }}
-            >
-              Sign out
-            </Button>
-          </Box>
-        </DialogContent>
-
-        <DialogActions>
-          <Button onClick={handleClose}>Close</Button>
-        </DialogActions>
-      </Dialog>
-    </>
-  );
-}
 
 export default function DrawerAppBar(props: Props) {
   const { window } = props;
@@ -103,18 +33,9 @@ export default function DrawerAppBar(props: Props) {
   const drawer = (
     <Box onClick={handleDrawerToggle} sx={{ textAlign: "center" }}>
       <Typography variant="h6" sx={{ my: 2 }}>
-        ⚡ Github Wiki
+        ⚡ Github API
       </Typography>
       <Divider />
-
-      <Button
-        variant="outlined"
-        startIcon={<LogoutIcon />}
-        onClick={() => signOut()}
-        sx={{ mt: 3 }}
-      >
-        Sign out
-      </Button>
     </Box>
   );
 
@@ -144,11 +65,11 @@ export default function DrawerAppBar(props: Props) {
               display: { xs: "none", sm: "block" },
             }}
           >
-            ⚡ Github Wiki
+            ⚡ Github API
           </Typography>
 
           <Box sx={{ display: { sm: "block" } }}>
-            <PopUpButton userData={session?.user} />
+            <ProfilePopup userData={session?.user} />
 
             <Button
               variant="outlined"
