+++ src/pages/index.tsx
@@ -1,19 +1,14 @@
 import React, { useEffect, useState } from "react";
 import { useSession, signIn } from "next-auth/react";
 import DrawerAppBar from "../components/navbar/navbar";
+import TableComponent from "../components/table/TableComponent";
+
 import {
   Avatar,
   Button,
   Typography,
   Container,
   Box,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Paper,
   TextField,
 } from "@mui/material";
 import GitHubIcon from "@mui/icons-material/GitHub";
@@ -147,66 +142,10 @@ const Login = () => {
 
           {searchResults.length > 0 && (
             <>
-              <TableContainer
-                component={Paper}
-                sx={{
-                  mt: 2,
-                  backgroundColor: "#191919",
-                  borderRadius: "10px",
-                  border: "1px solid gray",
-                }}
-              >
-                <Table sx={{ minWidth: 650 }} aria-label="simple table">
-                  <TableHead>
-                    <TableRow>
-                      <TableCell sx={{ color: "white" }}>Repo name</TableCell>
-                      <TableCell align="right" sx={{ color: "white" }}>
-                        Language
-                      </TableCell>
-                      <TableCell align="right" sx={{ color: "white" }}>
-                        Watchers&nbsp;
-                      </TableCell>
-                      <TableCell align="right" sx={{ color: "white" }}>
-                        Forks&nbsp;
-                      </TableCell>
-                      <TableCell align="right" sx={{ color: "white" }}>
-                        Id&nbsp;
-                      </TableCell>
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {searchResults.slice(0, displayedRepos).map((repo: any) => (
-                      <TableRow
-                        key={repo.name}
-                        sx={{
-                          "&:last-child td, &:last-child th": { border: 0 },
-                          color: "white",
-                        }}
-                      >
-                        <TableCell
-                          component="th"
-                          scope="row"
-                          sx={{ color: "white" }}
-                        >
-                          {repo.name}
-                        </TableCell>
-                        <TableCell align="right" sx={{ color: "white" }}>
-                          {repo.language}
-                        </TableCell>
-                        <TableCell align="right" sx={{ color: "white" }}>
-                          {repo.watchers}
-                        </TableCell>
-                        <TableCell align="right" sx={{ color: "white" }}>
-                          {repo.forks}
-                        </TableCell>
-                        <TableCell align="right" sx={{ color: "white" }}>
-                          {repo.id}
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </TableContainer>
+              <TableComponent
+                searchResults={searchResults}
+                displayedRepos={displayedRepos}
+              />
 
               {displayedRepos < searchResults.length && (
                 <Box sx={{ p: 2, textAlign: "center" }}>
